requirements.txt
streamlit
pandas
spacy
PyPDF2
python-docx
